test: cover the note handler seam and split handlers.test.ts under the cap
`note.file`'s persist-before-route order had no handler-level test, so the
branches that matter were unproven: an unrouted note is still durable, a
store failure degrades to routing-only instead of erroring the agent, and a
note never routes back to its own author. Adds those plus `note.list`, and a
launch test pinning that field notes seed worktree sessions but never the
main (dispatcher) session, which receives them live.

Touching handlers.test.ts put it at 605 lines, so task CRUD / issues /
worktree verbs move to a sibling suite and the shared fixtures move into the
existing context helper. No assertions changed.
